+++ libraries/runtime/src/nativeMain/kotlin/dev/ktool/embed/Resources.kt
@@ -18,7 +18,7 @@ import platform.posix.getenv
  */
 class Resources(
     private val resourceDirectory: ResourceDirectory,
-    private val inMemoryCutoff: Int = IN_MEMORY_CUT_OFF,
+    private val inMemoryCutoff: Long = IN_MEMORY_CUT_OFF,
 ) : BaseResources(
     resourceDirectory,
     inMemoryCutoff,
